Modify

+++ Source/Application.h
@@ -51,7 +51,7 @@ public:
 
 private:
 	// Frame time in ms
-	float frameTime = (1.0 / 60.0) * 1000;
+	uint frameTime = (1.0 / 60.0) * 1000;
 };
 
 // Global var made extern for Application ---
